// packages/shared/State.d.ts
export declare class McosEncryptionPair {
    _cipher: Cipher;
    _decipher: Decipher;
    /**
     * Create a new encryption pair.
     *
     * This function creates a new encryption pair. It is used to encrypt and
     * decrypt data sent to and from the client.
     *
     * @param {module:crypto.Cipher} cipher The cipher to use for encryption.
     * @param {module:crypto.Decipher} decipher The decipher to use for decryption.
     */
    constructor(cipher: Cipher, decipher: Decipher);
    /**
     * @param {Buffer} data The data to encrypt.
     * @returns {Buffer} The encrypted data.
     */
    encrypt(data: Buffer): Buffer;
    /**
     * @param {Buffer} data The data to decrypt.
     * @returns {Buffer} The decrypted data.
     */
    decrypt(data: Buffer): Buffer;
}
/**
 * The encryption settings for a session.
 */
export declare class McosEncryption {
    connectionId: string;
    _commandEncryptionPair: McosEncryptionPair;
    _dataEncryptionPair: McosEncryptionPair;
    /**
     * Create a new encryption object.
     *
     * @param {object} args
     * @param {string} args.connectionId The connection id of the session that
     *                                   this encryption is for.
     * @param {McosEncryptionPair} args.commandEncryptionPair The encryption
     *                                                        pair for
     *                                                       command packets.
     * @param {McosEncryptionPair} args.dataEncryptionPair The encryption pair
     *                                                    for data packets.
     */
    constructor({
        connectionId,
        commandEncryptionPair,
        dataEncryptionPair,
    }: {
        connectionId: string;
        commandEncryptionPair: McosEncryptionPair;
        dataEncryptionPair: McosEncryptionPair;
    });
    get commandEncryption(): McosEncryptionPair;
    get dataEncryption(): McosEncryptionPair;
}